LogOutHomePage UI is done
I have finished the UI needed for the Home page when the user is logged out which maps to the address: '/'. It looks almost identical to the real website page. The LogOutHomePage includes the components from the previous push. Most of these components will be also used in the LogInPage (/login) and the SignUpPage (/register) in later pushes.
I also did some basic layout for both pages. Added some routes and links.
Last Note:
The Divider in the LogOutHomePage appears black and for some reason changing it to white isn't easy, so I'll probably try to adjust it once I take care of the more important stuff.