Basic Login Using Google
As in the title, the basic Google LogIn functionality as a Pop Up window instead of redirect as in the original tumblr (I'll try to switch to redirecting later on but it needs more handling). I also made a tiny adjustment in the ProfileDropDown component for the sole sake of showing the Logged In State (for ex: showing the profile picture of the Google account). This can changed and modified later on.
I also haven't implemented the API calls (both mock and real).
N.B: I added a new package ( react-google-login ) so don't forget to run npm install.