Both Google SignUp & Login Done Using Real API
Handled cases 200 & 404 for login_with_google (and laid the bare bones for case 422).
Handled case 200  only for register_with_google. (422 & 500 to be handled later by putting it as a label).
If you want to register with Google, the first time you'd have to enter your blog name and age. Then after you log out and want to login again simply click on the ContinueWithGoogle button and you will be automatically logged in.