Removed the /logout path from all objects and replaced it with /
Added a login check: When the local host starts at path /, the user is automatically redirected to the dashboard if they are logged in and if not they are faced with the LogOutHomePage.
Added a brief check on the current path/location to view the correct buttons in the login and register pages as in the original tumblr.